Spike/add deploy schemas to instant seal node (#18)
Problem
=======
Testing with frequency instant seal node can require deploying schemas
to the node, which is an added layer of complexity when writing end to
end tests

Closes: [184003669](https://www.pivotaltracker.com/story/show/184003669)

Solution
========
Created a dockerfile in this repo to take latest frequency instant seal
node image and deploy schemas on top of it. Additionally added a github
actions workflow to update dockerhub with a new version and latest each
time we push a new tag. This will need to be manually done when
frequency updates versions to match with it.


Change summary:
---------------
Adding a docker image to deploy schemas to a frequency instant seal
node, which will aid with testing in test containers as well as improve
general quality of life when developing on frequency.

Steps to Verify:
----------------
To run locally
1. Make sure terminal is at schemas root directory
2. To build on intel mac (on a m1 chip switch "amd64" with "arm64"), run
this (and sub with whatever name you want the image to be: docker buildx
build --platform=linux/amd64 -f
docker/instant-seal-node-with-deployed-schemas.dockerfile -t .
3. To run image after building on intel mac, run this (and sub image
name with whatever you tagged the name in previous command): docker run
--rm -dp 9944:9944 -p 9933:9933 -p 30333:30333